requirements.txt
Django==3.2.16
django-filter>=23.2
djangorestframework>=3.14.0